cron/popular/requirements.txt
mysql-connector-python==8.0.33
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-sdk==1.22.0
psycopg2==2.9.9